refactor (LottoNumbers) : LottoNumbers로 다시 분리
